+++ src/org/apache/bcel/classfile/ConstantFloat.java
@@ -87,6 +87,7 @@ public final class ConstantFloat extends Constant implements ConstantObject {
     /**
      * Initialize from another object. Note that both objects use the same
      * references (shallow copy). Use clone() for a physical copy.
+     * @param c
      */
     public ConstantFloat(ConstantFloat c) {
         this(c.getBytes());
@@ -133,6 +134,7 @@ public final class ConstantFloat extends Constant implements ConstantObject {
 
     /**
      * @param bytes.
+     * @param bytes
      */
     public final void setBytes(float bytes) {
         this.bytes = bytes;
@@ -149,6 +151,6 @@ public final class ConstantFloat extends Constant implements ConstantObject {
      * @return Float object
      */
     public Object getConstantValue(ConstantPool cp) {
-        return bytes;
+        return Float.valueOf(bytes);
     }
 }
